+++ src/simulation/Simulation.h
@@ -18,8 +18,8 @@ class Simulation {
     const double simulation_end_time;
 
     const FileOutputHandler file_output_handler;
-    const size_t fps;
-    const size_t video_length;
+    const int fps;
+    const int video_length;
 
     const std::vector<std::unique_ptr<ForceSource>>& forces;
     std::unique_ptr<IntegrationFunctor> integration_functor;
@@ -39,7 +39,7 @@ class Simulation {
      * @param video_length Length of the simulation video in seconds
      */
     Simulation(ParticleContainer& particles, const std::vector<std::unique_ptr<ForceSource>>& forces, FileOutputHandler& file_output_handler, double delta_t,
-               double simulation_end_time, size_t fps = 24, size_t video_length = 30, IntegrationMethod integration_method = IntegrationMethod::VERLET);
+               double simulation_end_time, int fps = 24, int video_length = 30, IntegrationMethod integration_method = IntegrationMethod::VERLET);
 
     /**
      * @brief Runs the simulation, using the parameters given at construction
